Augment the RFC3164 parser to handle messages without a tag
A Vector user reported poor behavior when sending messages that had
a hostname but not tag: https://github.com/timberio/vector/issues/4811

Their example socket data seemed to indicate that there was an empty tag
due to the existence of two spaces between the hostname and the
message.

RFC3164 is a little vague on whether this would be a valid message:
https://tools.ietf.org/html/rfc3164#section-4.1.3

But RFC5424 does ducoment what a message lacking the app-name field
would look like (https://tools.ietf.org/html/rfc5424#section-6) which
would look like this, with two spaces in a row.

Signed-off-by: Jesse Szwedko <jesse@szwedko.me>