Measure cycles per byte, introduce more benchmarks
This commit introduces new benchmarks. The work done in the previous commit
suggested that the parser was abnormally slow, about 200Mb/s, and with time
dominated by the `structured_data` parse. This is not wrong but not a complete
picture. This commit introduces further commits with longer lines, demonstrating
that the longer the line the better the throughput. This suggests that there is
a smallish overhead for parsing each line -- presumed to be allocation for the
`Message` structure but not demonstrated -- and that this cost is paid off
quickly once the line length increases.

Commit also introduces measurement of these benchmarks by cycle count, rather
than wall time. This is more stable, though not as stable as
[iai](https://github.com/bheisler/iai) will be once it's integrated into the
criterion toolset.

As it stands, short-lines cost about 10 cycles per byte to parse, long-lines
cost 1 cycle per byte, rounding up some. If we need to improve this in the
future we ought to explore what the overhead per line is. One avenue for that is
to adjust the interface to accept a `&mut Message` to populate, rather than
allowing it to alloc the `Message` itself.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>